Refactor code, add API for setting visibility
The current alert message API allows a message to be shown only once, when it is attached
to the DOM. If the user dismisses the alert, or the alert is auto dismissed, it cannot
be shown again without being detached from the DOM and re-attached.

This commit adds functionality to the public API to dynamically show and hide the alert
message on demand via the `visible` property. It also exposes some additional properties
to allow the developer to expand and collapse the alert message via. the `expand`
property, and to know when the alert message is overset via. the `overset` read-only
property.

To maintain backwards compatability, the alert message `visible` property will be set
to true automatically when it is attached. (Developers using the current API expect
that to happen.) The `disableAutoShow` property has been added to allow developers
to turn off this behavior. This will be useful in apps where alert messages are
stamped out in advance and shown when needed, then hidden and re-used for another
message later.